refactor: app 모듈 정리 (#46)
